Use getDebugPositionAtOffset only when it is needed
- getDebugPositionAtOffset is costly due to the search of
LineNumberNode. Into JarSourceCode, use this method only for
the first instruction of a block because after current position
is automatically updated by LineNumberNode contained by the
current block without need to recompute this information
When compiling framework, the number of calls to this method
is divided by 5.9 and allow to remove this method from hotspot.

Bug: 74712202
Change-Id: Ie676471f9d8918466ca985689cb7803919d20f29