Make forwarding methods synthetic when desugaring
When desugaring a default method of an interface, we create a
forwarding method in every "implementation" class. This forwarding
method then calls into the companion class that contains the code of
the default method.

If a debugger is attached to a program that has been desugared and
steps into the call to the default method, it will actually suspend
in the forwarding method. Thus it requires an extra single step to
reach the code of the default method (in the companion class).

By making the forward method synthetic, we allow debuggers (like
IntelliJ) to automatically skip such method during single step.
Therefore we better match the behavior when running the same program
without desugaring by reaching the code of the default method in only
one step.

Bug: 77634519
Change-Id: I1b9879509bd21e461ac249ec11638920fce4f26a