Find more phis that have known-to-be-non-null values.
Currently, after adding a non-null IR, phi users, whose blocks are
dominated by the block with that non-null IR, are identified as
dominated phi users. This is conservative in that we only detected
phis only if all of its known-to-be-non-null operands are dominated,
e.g., phi(v0, v1, v1) -> phi(v0, vn, vn) where phi's block is dominated
by the block with vn.

We should consider incoming edges, instead of phi's block:
if the operand is non-null, and the corresponding predecessor is
dominated by the block with non-null IR, we can selectively replace
that operand with known-to-be-non-null value,
e.g., phi(v0, v1, v1) -> phi(v0, vn, v1) where the 2nd predecessor is
dominated by the block with vn.

Refer to http://b/76202537#comment4 and http://b/76202537#comment5 for
more detailed discussion and examples.

Bug: 76202537
Change-Id: I39e05bfe3ff070b1fdf83f89d528c7facae8973c