Warn when lambda implements unknown library interface
In the CF backend, we need all interfaces implemented by lambdas to be
in the input in order to ensure proper renaming of interface methods.
If R8 is run with minification on an incomplete program with a
multi-interface lambda implementing both a missing library interface and
a present program interface, then we might rename the program interface
method even though it should be kept the same as the (missing) library
interface method.

This issue was uncovered by the recent buildbot failure that was fixed
in I9f42d46b b530d5f41 ("Fix LambdaTestRunner using Android jar for Java
Runtime test", 2018-04-24). Fixes the incorrect handling introduced in
Id187bb9a 8492920fb ("Keep track of instantiated lambdas", 2018-04-20).

Change-Id: Ib94422ccd5746159bce60a21402be8a2389c86af