Debugging tests: fix breakpoint setup when line table is empty
When we set a breakpoint on the first line of a method, we request
its line table then select the first line in this table and set the
breakpoint on the corresponding instruction index.

If a method has no line table (which is the case for synthesized
methods), then we just need to set the breakpoint on the first
instruction.

Change-Id: I9d00f06d22bc5f2c604d41199710c569662f7c53