Change dex partition to reduce LinearAlloc usage
Attempt to partition classes among dex files to limit linking errors
during DexOpt and when it's not possible to avoid then, ensure that they
won't cause LinearAlloc over usage.

This is achieved by ensuring that each class in a secondary dex has all
it's hierachy (super classes and interfaces) either in the bootclasspath
in the same secondary dex. Then if it's not possible to respect this
constraint for some classes, instead ensure that those classes are put in
a different secondary dex than all their link dependents (i.e. subclasses,
implementations or sub interfaces), those classes will failed to link
during DexOpt but they will be loaded only once.

Bug: 73108937
Change-Id: Ib4af372732232d20814ebaad2e2265c99622de6a