Transform check-casts prior to reg.alloc.
This CL rewrites casts on the form "lhs = (T) rhs" into "(T) rhs" in selected bridge methods. This transformation helps to ensure that we do not insert unnecessary moves in bridge methods with an invoke-range instruction, since all the arguments to the invoke-range instruction will be original, consecutive arguments of the enclosing method (and importantly, not values that have been defined by a check-cast instruction).

Change-Id: Ief382c386d9654dbe84877f991e32b8dd60186f5