CF backend: Preserve broken parameter annotation attribute
When ASM reads in a classfile with a parameter annotation attribute that
has fewer parameter annotations than the actual parameter list, it emits
special "java.lang.Synthetic" annotations on the parameters that do not
have entries in the parameter annotation attribute.

Store the number of Synthetic annotations seen in DexAnnotationSetRefList
and use it in CfApplicationWriter to re-emit Synthetic annotations to
ASM so that it can maintain the broken attributes.

Change-Id: Id9d602438ceb73f70cb80be0bf20215c611176f8