CfSwitch: Implement table switches
Implement support for reading table switches in the testing CF frontend
and in the CF backend. Note that IR->CF construction does not actually
emit table switches.

Also disable collection of switch maps and ordinal maps in R8 when
skipIR is set. This is required for CfFrontendExamplesTest in the
"switchmaps" example until CF->IR construction is complete.

Change-Id: Id38dbb01634676c79d2ea559ab98dfb71d051735