Simplify error reporting.
This CL proposes that we ensure that origins are the defining place for
locations of resources. For example, API methods on paths are changed to encode
the context of the inputs in the origin. We assume the origin to have this
information and thus, IOExceptionDiagnostics should not attempt to retrive it
again. We should, in time, avoid throwing IOException in the compiler and
instead associate proper origins at the point of exception. This is not
completed in the present CL.

Bug: 78893028
Change-Id: I750e3837a70799a8ee44b51a486a70bf20791319