Make error reporter accessible to internal AndroidApp container.
This is a prerequisite to cleaning up error messages as most IOExceptions
originate in the AndroidApp API.

In addition to defining the reporter in AndroidApp, the builders have been
cleaned up and aligned. The default mode initialization is also shared now.
Note the previous code has an unnoticed bug setting R8's default mode to debug
if a diagnostics handler was passed.

Bug: 78893028
Change-Id: Ib372c4b6f4d0f044cd2936ccccec1bc7fca25d78