Make CF frontend for testing lazy
Move JarCode.ReparseContext to JarClassFileReader.

Introduce new class LazyCfCode which implements isCfCode() and
asCfCode() pretending to be a CfCode, and doing the lazy parsing when
asCfCode() is invoked the first time.

Rename and move JarClassFileReader.CfCreateMethodVisitor to
LazyCfCode.MethodCodeVisitor.

Also use ASM's JSRInlinerAdapter to remove the JSR instruction before
parsing with LazyCfCode.MethodCodeVisitor.

Change-Id: I345b74c14caf3aa525e3a45b653ec1f4ed52c3fa