Add support for -if rule with back references.
Introduced (ProguardWildcard), which has two subtypes:
  <1>.Pattern, such as *, **, ?, and
  <1>.BackReference, which has a (back) reference to Pattern.

While building -if rule, along with its subsequent rule, back references
are linked to target pattern.

In *Matcher#matches, the captured part for the pattern is recorded, which
is in turn used to eat the counter part of linked back reference.

With the presence of back references, RootSetBuilder is re-implemented to
take into account
  1) all pairs of -if rule and live types, and
  2) all combinations of live members,
because, depending on which types and members trigger the -if rule, the
application of the subsequent -keep rule may vary (e.g., due to <n>).

Bug: 73800755, 79486261
Change-Id: I019d29726737a050f90540c66838a5491f5a5a03