Make output deterministic for dex code with invoke-custom.
DexCallSites and DexMethodHandles were not sorted. Sorting is not
needed for correct code, but it is needed for deterministic output.

The complication for DexCallSites is that two call sites can be
equal and only differ in identity. Therefore, we have to have
some external thing to sort based on. This change uses the
method and instruction offset of the invoke-custom in the output
that the call site is needed for to deterministically sort them.

Bug: 78569026
Change-Id: I4b9e8f36b533f4dd1eb288753dbefdbfb08ab464