ExamplesDebugTest: Extend with many more examples
While most examples are run on both JVM and Art with the output of both
the CF and DEX backends, some tests only run on JVM with CF backend and
some are disabled altogether.

Introduce getCfConfig() which takes an InternalOptions Consumer
in anticipation of adding testing of the new CF frontend when CF->IR
construction is implemented.

Change-Id: I28473b50f48e4e48f891ffb6e4f4880554ba875e
Bug: 79671093