Add command-line program JarDiff
Add a program to compare two JARs. Given two JARs as input, the program
first outputs a list of classes only in one of the input JARs. Then, for
each common class, the program outputs a list of methods only in one of
the input JARs. For each common method, the output of CfInstruction
.toString() on each instruction is compared to find instruction-level
differences. A simple diffing algorithm is used that simply removes the
common prefix and common suffix and prints everything from the first
difference to the last difference in the method code.

Change-Id: Ibb9f0ddee5e5266887c14dbeefa9b00f598e498f