Refactor and rename DexAnnotationSetRefList
Rename the class to ParameterAnnotationsList since the class is only
used for representing parameter annotations. Make `values` private and
replace it with a method that visits all the contained DexAnnotations.

Due to a javac bug that went unfixed for multiple Java versions, the JVM
specification does not require that the number of entries in the
ParameterAnnotations attribute of a method matches the number of
parameters in the method prototype; the number of ParameterAnnotations
entries may be less than the number of prototype parameters for methods
on inner classes.

Change DexParser to pass the actual number of prototype parameters to
ParameterAnnotationsList (in case parameters are missing annotations).

Make FileWriter use countNonMissing() and isMissing() to maintain the
broken ParameterAnnotations attribute created by the javac bug.
The JAR frontend and CF backend preserve the brokenness as they have
done since Id9d60243 32f87feeb, "CF backend: Preserve broken parameter
annotation attribute", 2018-05-03.

There are two ways of accessing the parameter annotations:
* Iterating over the list using forEachAnnotation()
* Using size(), get(i) and isMissing(i)

The forEachAnnotation() method visits all the DexAnnotations specified
in the ParameterAnnotations attribute.
In contrast, the size() and get(i) methods may be used to access the
annotations on individual parameters; these methods automatically shift
parameter annotations up to mitigate the javac bug. The isMissing(i)
accessor is used to determine whether the i'th parameter is missing in
the ParameterAnnotations attribute.

Also fix keepIf() throwing some parameter annotations away: For a method
with annotations on two parameters, if the predicate passed to keepIf()
throws away some annotations on the first and some on the second
parameter, then all of the annotations on the second would be kept
anyway. (Broken since the code was added in I6d6b2d04 19da09429
"Implement annotation removal.", 2017-03-10, in the old prerelease
repository.)

Change-Id: Ib47292a9c65b6fc696f6a5b922bfff7e467bbeb5