Tests for processing of CovariantReturnType annotation
This CL adds a set of initial tests for the (upcoming) transformation of dalvik.annotation.codegen.CovariantReturnType (and dalvik.annotation.codegen.CovariantReturnType$CovariantReturnTypes).

The tests rely on three versions of the same library. Version 2 of the library uses the @CovariantReturnType annotation. The idea is that D8 should generate synthetic methods with the given (covariant) return type that simply forward to the method that the annotation has been attached to. Then, in version 3 of the library, these @CovariantReturnType annotations have been removed, and the signatures of the methods have been updated. As a result, any client that used to work with version 2 of the library will also work with version 3 of the library.

Bug: 78618808
Change-Id: Iaa7b7460438e139b279e78f0c271854c7c35635c