Only use JSRInlinerAdapter if necessary
Since JSR and RET instructions are not emitted by modern Java compilers,
they will almost never be encountered in practice. Try parsing input
classfiles without JSRInlinerAdapter; if JSR or RET is encountered,
restart parsing the problematic class with the JSRInlinerAdapter.

In JarCode, this is implemented by iterating over the instructions,
looking for JSR and RET explicitly.

In LazyCfCode, a private exception class is thrown to restart processing
the class as soon as JSR or RET is encountered.

JumpSubroutineTests: Add testing of CF frontend.

CfStackInstruction: Enable building IR for the SWAP instruction,
since this is encountered in JumpSubroutineTests.
Change-Id: Ic6865d1e11ecb10b1388dded3b90fcfca4f18d7d