Adjustments from primitive types to Comparable/Serializable
In method reference desugaring we need to adjust types of
parameters and return values. One of such adjustments allows
conversion from primitive type to correspondent boxed type
or any of its supertypes.

Thus, to properly calculate allowed adjustments we need a
classpath and, since we want to avoid it for lambda desugaring,
the current implementation only allowed primitive type adjustments
to correspondent boxed type, java.lang.Object, and in appropriate
cases to java.lang.Number.

With this CL we are adding java.lang.Comparable and
java.io.Serializable to this list to match Java 8 class
hierarchy (still without requiring classpath).

Bug: 79986240
Change-Id: I1bb21621ac8abaf1a306627d52d0fa22abb06c20