Processing of @CovariantReturnType annotations
This CL is responsible for processing the annotations dalvik.annotation.codegen.CovariantReturnType and dalvik.annotation.codegen.CovariantReturnType$CovariantReturnTypes. If such an annotation is attached to a method m, then the compiler should insert a new synthetic method that is equivalent to method m, but has the return type specified by the annotation.

Bug: 78618808
Change-Id: Id4ea6c1e3f5c97a57954b1ededca473a022736a3