R8 Kotlin tests: also check symbols in input
Some of the test for Kotlin verify that R8 is able to remove symbols
(classes, methods or fields) generated by Kotlin compiler. However we
used to never verify that the symbol exists in the input. This could
lead to some test being successful despite being incorrect.

This CL fixes that issue by also checking that the expected symbol
exists in the input classpath. It also takes the opportunity to rename
some utility methods to avoid confusion. For example, the checks for
method symbols have been renamed to:
* checkMethodIsAbsent: checks that the method is absent both in the
input and the output.
* checkMethodIsKept: checks that the method exists both in the input and
the output.
* checkMethodIsRemoved: checks that the method exists in the input but
no longer exists in the output.
The same logic is applied for classes and fields.

This CL also updates some tests that were incorrect (detected by this
change).

Change-Id: Ia55ac5d7049cd38f16d8742593ae0a5d68c977d8