Add simple flag file support to D8 and R8.
This allows to create a file containing the arguments for the
invocation. This can be helpful when the command-lines get too
long for the system to deal with.

The support is very simple. In particular, since there are spaces
between the option name and the value they have to be on separate
lines with this version:

d8 --output out.zip input.jar

becomes

d8 @flags.txt

where flags.txt contains:

--output
out.zip
input.jar

If this turns out to be problematic we can change it to split
lines. However, that gets complicated if there is white-space
in files names.

R=sgjesse@google.com, zerny@google.com

Bug: 79985374
Change-Id: Ic439fa9873511086eda8ef5a6a937c664106584b