Make ClassNameMinifier resilient to the absence of (shrunk) outer class.
When renaming an inner class to keep its signature and inner attribute,
the minifier tried to search for the outer class and the corresponding
naming space. If that outer class is shrunk, we've used the outer class
name as the prefix, which could be not consistent with the package
renaming scheme. E.g., if all the other classes in the package is
remapped to or flatten to the top-level, keeping inner classes in the
original package (via the shrunk outer class's binary name) caused an
illegal access error.

Although outer classes would be shrunk, we can still use their types to
pretend they are there and generate naming spaces for them, which just
help us assign consistent renames to inner classes.

Bug: 80083341
Change-Id: I40ef60f7dd1360c1265393f5963ac580189ad1e1