Mark field type as live.
We assume that a field type is equal to, or at least a super type of,
the type used to initialize that field. If the type relation is broken
for some reasons, that's not the case anymore. It is safer to mark the
type of field as live when we mark the field as live.

Bug: 78788577, 80206476, 80130830
Change-Id: I114601eb02d6b8f0802cd50805beb07a637266df