Make ASM skip code when reading class structures
Pass the SKIP_CODE flag to ASM's ClassReader to make ASM skip decoding
the opcodes of each method's code on the first pass through the input
classes where we only build up the class structure of the program.
Since our subclass of ClassVisitor does not implement any of the
instruction visit methods, there is no reason for ASM to decode
instructions and call the default empty visit*() implementations.

Change-Id: I10f4f9b3dd4d7c3c2201bfd879c367e0812edbae