Specify API surface of DexSplitter
DexSplitter and FeatureClassMapping are public API and should be
annotated with @Keep and marked final.

DexSplitter.Options is public API, but DexSplitter.FeatureJar is not.
As a result, make addFeatureJar(FeatureJar) private and getFeatureJars()
package-private.

FeatureClassMapping.fromJarFiles() takes a List<FeatureJar> and is
needed from a different package of R8, but is not public API, and
therefore it is placed in a static inner FeatureClassMapping.Internal
class (which is not annotated with @Keep).

Also replace DexSplitter.Reporter by DefaultDiagnosticsHandler.

Change-Id: I7fc53083968763de285fadb83f2f50b8f7c479da