Support --pg-conf in Disassemble for class-file input
Implement toString(DexEncodedMethod, ClassNameMapper) on CfCode and
LazyCfCode by passing the ClassNameMapper to CfPrinter and using the
mapper in all the right places.

Make Disassemble use the CF frontend for testing to ensure that
LazyCfCode is used to disassemble instead of JarCode (for which it would
be difficult to support ClassNameMapper).

Also ensure that the output of Disassemble with --pg-conf is similar to
the output without.

Change-Id: Ia36dd7f17515be330de2da5b34d40583855a4feb