Test for "R8/CF does not merge identical catch handlers"
Also add ToolHelper.runR8WithConsumer(AndroidApp, ProgramConsumer) to
support compiling an AndroidApp with the CF backend.

Unfortunately the name runR8(AndroidApp, ProgramConsumer) clashes with
runR8(AndroidApp, Consumer) because both Consumer and ProgramConsumer
are functional interfaces, and runR8(AndroidApp, Consumer) is used in
many places with a lambda as second argument.

Bug: 80514966
Change-Id: I15e88849527285ea45248f8c0ab8fac1e30e84fb