Regress78493232_WithPhi: Fix expectation for failing test
The test introduced in Ic278219f9 72548942a ("Remove trivial phis for
all string new-instance values.", 2018-06-01) only checks that the
stdout when running on ART is the same as when running on JVM.

Since the test program fails with empty stdout on JVM and all ART except
for version 6.0.1, the test would only fail (different stdout) when run
on Art 6.0.1. However, the behavior is only the same on ART 5.1.1 (the
test method returns null); on all other ART versions, the behavior is
different from the JVM:

* On 4.0.4, 4.4.4, 7.0.0 and tip-of-tree, the test program is rejected
  by the Dalvik/ART verifier after being processed by D8.

* On 6.0.1 a different execution is observed:
  "java.security.SecureRandom" is returned by the test method.

Ensure that the test does not throw NullPointerException on JVM and ART
5.1.1 by swapping a null and non-null argument to .equals(), and specify
the observed incorrect outcomes instead of using ensureSameOutput().

Also assert in AsmTestBase.ensureSameOutput() that exitCode is 0.

Bug: 78493232, 80118070, 109645295
Change-Id: I06392e5c0ec34de26cb154c4ce7178ad84f3ebe0