Do not translate File.separatorChar to / when reading from archives.
We already have the split between Paths and Strings when guessing
descriptors. Consistently use the String based on for archive
entry names and the Path based one for file names.

This will make sure that an archive provider will not accidentally
include a descriptor that cannot be read from the archive.

R=gavra@google.com, sgjesse@google.com

Bug: 80510937
Change-Id: Ia895820cb260aabaf091be830077d5d2e1b53191