Extend context sensitive GraphLense API
Extends the method GraphLense.lookupMethod with an argument of type Invoke.Type, such that graph lenses will have to opportunity to be context sensitive in the type of the invoke. This is needed for vertical class merging.

In addition to the lookupMethod API extension, this CL adds a new GraphLense (VerticalClassMergerGraphLense) that will be instantiated by the vertical class merger.

Change-Id: Ic6b72eacbb0f2cce8b9cda24a9d92722784bc5f3