Eliminate redundant field loads.
Field loads that are dominated by a load of the same field on the same
object and for which there are no locking operations, stores to that
field, stores to static fields of other objects, or calls in between
the last load and this load can be eliminated.

This saves ~50kB on framework code.

Change-Id: I42853a6caf67e3e6caa47a0eac1227519b6a8490