Rewrite super calls in vertical class merging
This CL handles class merging in presence of invoke-super calls. If a class A gets merged into a class B, and a method in B invokes a method in A via an invoke-super instruction, then that invoke-super instruction needs to be changed into an invoke-direct instruction after class merging (an invoke-virtual instruction would potentially hit an implementation in a sub class of B, which would change the semantics).

To do this, this CL updates the vertical class merging such that all virtual methods in the class A are converted into direct methods in B. In addition, a virtual bridge method is added to B if the given method from A is not shadowed in B (such that the method is not removed from the interface of B).

Change-Id: I22650d8f4e59ee98266b2073262e2ea97c71dc0d