Handle non-standard invoke-super in class merging
This CL handles rewriting of invoke-super instructions that would not normally be generated by javac. Necessary for InvokeSuperTest to succeed with vertical class merging enabled.

The issue handled by this CL is the following.

If we merge class C into class D and class D contains an invocation super.m(), then it is insufficient to rewrite all instances of "invoke-super C.m()" in D to "invoke-direct D.m$C()" (the method D.m$C denotes the direct method that has been created in D for C.m). In particular, there might be an instruction "invoke-super B.m()" in D that resolves to C.m at runtime (assume that B is a superclass of C), which also needs to be rewritten to "invoke-direct D.m$C()".

Note that D might also have an instruction "invoke-super A.m()" that would lead to a NoSuchMethodError at runtime (A is a superclass of B). In this case, it is important not to rewrite this instruction into "invoke-direct D.m$C()", since this would change the semantics.

Change-Id: I399fc5a5e26954d71416915b8aaadd6405737e29