Pin signatures of pinned fields/methods in class merger
If a field or method is pinned, then none of the types mentioned in the signature of that field or method should be considered for vertical class merging.

The test ProxiesTest explicitly keeps a method "private void x(com.android.tools.r8.shaking.proxy.testclasses.BaseInterface)" and checks that a method with this signature is present in the output. Prior to this CL, vertical class merging would merge BaseInterface into its sub class, causing this test to fail.

Change-Id: Id8d8633e4e7311d740394ad6ea5e2150d504cc5b