Don't check the current line when popping an activation frame.
Due to the usual placement of a store after an invoke, stepping out of a call
will generally place the debugger back on the line of the call. This is however,
not always the case as the stack operations may be removed, eg, the
new/dup/init/astore pattern may be different. In general, we don't care from a
debugging perspective if the step-out goes to the line of the call, or to the
line of the instruction following the call.

Bug: 79671093
Change-Id: I73263cda0320b3e23897fc20e4ccc176e2c25c28