Debugging states should be filtered when computing the 'next' state.
Prior to this we would still need to have the same number of states for each
stream, just a filtered-out state would be allowed to differ among each stream.
With this change, filtered states are simply discarded from the streams and we
always compare the first non-skipped points of the streams with each other.

Change-Id: I54668d0a54b4f3c2de3ecc6ef877ebe120637de4