Move invocation type mapping to GraphLense
Method lookup in a GraphLense now returns both the method to invoke and
the invocation type to use as its result.

Before this happened in the LenseCodeRewriter after mapping the
method. This was somewhat fragile and disconnected from the method
mappping. Each lense providing the method mapping should also have the
information to determine the invocation type to use for the mapped
method.

Change-Id: I4dc65f5df8bfd5a392d8f660b83c61f7acd35282