Fixing inliningConstraintForVirtualInvoke for cases when no targets were found
Apparently there is something wrong with lookupTargets(...) called on
interface invocation. None of testInterface* methods in added tests are
inlined because we failed to find targets. Even such straightforward one
as testInterfaceD.

Note that not finding targets for interface invocations is not a big
problem, since in valid code all interface method implementations are
required to be public, so the check for targets being accessible (below)
should always succeed unless the input is erroneous.

Bug:
Change-Id: Ia5fcac3aa026739c73ff8fa1c4deac19c0ae7df0