Timing for collision detection in class merger
This CL also prints timing for the collision detection in the vertical class merging if options.printTimes is enabled.

To accomplish this, this CL extends the Timing class, such that it is possible to restart an existing timer. This is needed since the collission detection runs as one out of many checks inside a loop.

Change-Id: I8777d80d20f102072dc16634c8febc83ec3c1a70